+++ src/bootstrap.php
@@ -18,7 +18,7 @@ class SanitizerData
      * 
      * @var array
      */
-    public static $config = array(
+    const config = array(
         "maxInputLength" => 100, // 1
         "encoding" => "UTF-8", // 2
         "preventXSS" => true, // 3